Refactor: put helper functions in util.rs
