+++ src/sm4/mod.rs
@@ -21,6 +21,8 @@
 //! Most variable's name are same as those in the Document written by the Encryption Administration.
 //!
 //! [DACAS: sm4](http://dacas.cn/sharedimages/ARTICLES/SMAlgorithms/SM4.pdf)
+use basic::util::{bytes_to_four_u32, four_u32_to_bytes};
+
 pub const BLOCK_SIZE: usize = 16;
 pub const KEY_SIZE: usize = 16;
 pub const RND_KEY_SIZE: u32 = 32 * 4;
@@ -245,48 +247,6 @@ macro_rules! sm4_rnds {
         $B2 ^= sm4_t!($B0 ^ $B1 ^ $B3 ^ $Rk[$k2]);
         $B3 ^= sm4_t!($B0 ^ $B1 ^ $B2 ^ $Rk[$k3]);
     };
-}
-
-fn bytes_to_four_u32(b: &[u8; 16]) -> [u32; 4] {
-    [
-        (u32::from(b[0]) << 24)
-            + (u32::from(b[1]) << 16)
-            + (u32::from(b[2]) << 8)
-            + u32::from(b[3]),
-        (u32::from(b[4]) << 24)
-            + (u32::from(b[5]) << 16)
-            + (u32::from(b[6]) << 8)
-            + u32::from(b[7]),
-        (u32::from(b[8]) << 24)
-            + (u32::from(b[9]) << 16)
-            + (u32::from(b[10]) << 8)
-            + u32::from(b[11]),
-        (u32::from(b[12]) << 24)
-            + (u32::from(b[13]) << 16)
-            + (u32::from(b[14]) << 8)
-            + u32::from(b[15]),
-    ]
-}
-
-fn four_u32_to_bytes(l: &[u32; 4]) -> [u8; 16] {
-    [
-        (l[0] >> 24) as u8,
-        (l[0] >> 16) as u8,
-        (l[0] >> 8) as u8,
-        l[0] as u8,
-        (l[1] >> 24) as u8,
-        (l[1] >> 16) as u8,
-        (l[1] >> 8) as u8,
-        l[1] as u8,
-        (l[2] >> 24) as u8,
-        (l[2] >> 16) as u8,
-        (l[2] >> 8) as u8,
-        l[2] as u8,
-        (l[3] >> 24) as u8,
-        (l[3] >> 16) as u8,
-        (l[3] >> 8) as u8,
-        l[3] as u8,
-    ]
 }
 
 /// Compute round keys from a given key.
